test(service_test.py): updated tests to verify newletter content
